Add levelBounds and fix sprite flash glitch.
Units and weapons now handle level boundaries based on _levelBounds
attribute of each level (which can be assigned by defining LevelWidth and
LevelHeight as attributes under a Level element in LevelData.xml).
This should set the stage for implementing levels larger than the screen
and a camera class.

Sprites now flash properly and set color to grey when disabled.