Add composite ParticleEffects.
The old "ParticleEffect" is now called a ParticleGenerator.
The new "ParticleEffect" is a composite of 1 or more ParticleGenerators.

Instantiate a ParticleGenerator just like a ParticleEffect by passing in a
string that references an entry in ParticleEffects.xml.
If an entry for such an effect exists, it will instantiate an effect
containing all the defined generators.
If no such ParticleEffect exists, it will search for a ParticleGenerator
with the same name, and instantiate a ParticleEffect with only that one
generator (in other words, all the old particle effects now work as a
ParticleEffect wrapped around a single generator).