Improve flexibility of particle effect creation
A particle effect is, as before, defined as a collection of one or more
particle generators. However, one may now override individual attributes
of particle generators in their instantiation in a particle effect.

For example, the particle effect xml element:

<ParticleEffectData Name="MyParticleEffect">
	<ParticleGenerator Name="Effect1" StartColor="255,255,0,0"/>
	<ParticleGenerator Name="Effect1" StartColor="255,0,255,0"/>
	<ParticleGenerator Name="Effect1" StartColor="255,0,0,255"/>
</ParticleEffectData>

creates a particle effect consisting of three instances of Effect1, each
with a different start color.